batch: Move ownership unit rebuild
Ownership unit construction now has separate selection helpers while
callers still rebuild batch metadata after filtering or previewing unit
records.

The rebuild operation still lives in the construction module. That points
metadata reconstruction back at display grouping even though rebuilding
only consumes completed units and ownership claim formatting helpers.

This commit continues separating ownership unit responsibilities by moving
metadata rebuild into an ownership unit rebuild module and updating source
callers to import that boundary.

Subsequent commits will update the ownership unit tests and architecture
guards for the rebuild owner.
