+++ src/git_stage_batch/batch/ownership_units.py
@@ -10,13 +10,10 @@ from ..i18n import _
 from .display import build_display_lines_from_batch_source_lines
 from .ownership import (
     BatchOwnership,
-    ReplacementUnit,
 )
 from .ownership_claims import (
     LineRangeBuilder,
-    format_ownership_line_set,
     parse_ownership_line_ranges,
-    presence_claims_from_source_lines,
 )
 from .ownership_unit_types import (
     OwnershipUnit as _UnitRecord,
@@ -384,45 +381,6 @@ def validate_ownership_units(units: list[_UnitRecord]) -> None:
                     raise MergeError(
                         _("Invalid deletion in batch metadata: expected removed lines only.")
                     )
-
-
-def rebuild_ownership_from_units(units: list[_UnitRecord]) -> BatchOwnership:
-    """Rebuild BatchOwnership from semantic ownership units.
-
-    Args:
-        units: List of ownership units to combine
-
-    Returns:
-        New BatchOwnership with combined ownership from all units
-    """
-    all_presence_lines = LineRanges.empty()
-    all_deletions = []
-    replacement_units: list[ReplacementUnit] = []
-
-    for unit in units:
-        all_presence_lines = all_presence_lines.union(unit.claimed_source_lines)
-        deletion_indices = []
-        for deletion in unit.deletion_claims:
-            all_deletions.append(deletion)
-            deletion_indices.append(len(all_deletions) - 1)
-        if (
-            unit.kind == _UnitKind.REPLACEMENT
-            and unit.preserves_replacement_unit
-        ):
-            replacement_units.append(ReplacementUnit(
-                presence_lines=format_ownership_line_set(unit.claimed_source_lines),
-                deletion_indices=deletion_indices,
-                origin=unit.replacement_origin,
-            ))
-
-    return BatchOwnership(
-        presence_claims=presence_claims_from_source_lines(all_presence_lines),
-        deletions=all_deletions,
-        replacement_units=normalize_replacement_units(
-            replacement_units,
-            deletion_count=len(all_deletions),
-        ),
-    )
 
 
 def build_ownership_units_from_batch_source_lines(
