+++ src/git_stage_batch/batch/selection.py
@@ -10,8 +10,8 @@ from .ownership import BatchOwnership
 from .ownership_units import (
     build_ownership_units_from_batch_source_lines,
     validate_ownership_units,
-    rebuild_ownership_from_units,
 )
+from .ownership_unit_rebuild import rebuild_ownership_from_units
 from .ownership_unit_selection import select_ownership_units_by_display_ids
 from ..core.line_selection import (
     LineRanges,
